+++ controllers/user.js
@@ -1,15 +1,19 @@
+const Book = require('../models/book');
 const User = require('../models/user');
 const { validationResult } = require('express-validator');
 const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
 
 //Fetching Users
 exports.getUsers = async (req, res, next) => {
     try {
         const users = await User.find();
-        res.status(200).json({
-            message: 'Fetched Users Successfully!!',
-            users: users
-        });
+        res
+            .status(200)
+            .json({
+                message: 'Fetched Users Successfully!!',
+                users: users
+            });
     }
     catch (err) {
         if (!err.statusCode) {
@@ -23,7 +27,11 @@ exports.getUsers = async (req, res, next) => {
 exports.signUp = async (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        return res.status(422).json({message :errors.array()[0].msg})
+        return res
+            .status(422)
+            .json({
+                message: errors.array()[0].msg
+            })
     }
     const name = req.body.name;
     const email = req.body.email;
@@ -31,7 +39,11 @@ exports.signUp = async (req, res, next) => {
     try {
         const exUser = await User.findOne({ email: email });
         if (exUser) {
-            return res.status(200).json({ message: 'User with this email-id exists. Take Please try with different email-id' })
+            return res
+                .status(422)
+                .json({
+                    message: 'User with this email-id exists. Take Please try with different email-id'
+                })
         }
 
         const hashedPw = await bcrypt.hash(password, 12);
@@ -44,12 +56,14 @@ exports.signUp = async (req, res, next) => {
             fav: { books: [] }
         });
         await user.save();
-        res.status(201).json({
-            message: 'User created successfully!',
-        });
+        res.
+            status(201)
+            .json({
+                message: 'User created successfully!',
+                user: user
+            });
 
     }
-
     catch (err) {
         if (!err.statusCode) {
             err.statusCode = 500;
@@ -59,17 +73,74 @@ exports.signUp = async (req, res, next) => {
 
 };
 
+//Authorization
+exports.login = async (req, res, next) => {
+    const email = req.body.email;
+    const password = req.body.password;
+    let currentUser;
+    try {
+        const user = await User.findOne({ email: email });
+        if (!user) {
+            return res
+                .status(422)
+                .json({
+                    message: 'User with this email Not Found! Try signing up'
+                })
+
+        }
+        currentUser = user;
+        const isEqual = await bcrypt.compare(password, user.password);
+        if (!isEqual) {
+            return res
+                .status(422)
+                .json({
+                    message: 'Password incorrect'
+                })
+        }
+        const token = jwt.sign(
+            {
+                email: currentUser.email,
+                userId: currentUser._id.toString(),
+                role: currentUser.role,
+                userAccess: currentUser.userAccess
+            },
+            'libapp005567',
+            { expiresIn: '1h' }
+        );
+        res
+            .status(200)
+            .json({
+                token: token,
+                userId: currentUser._id.toString(),
+                role: currentUser.role,
+                userAccess: currentUser.userAccess
+            });
+    } catch (err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }
+};
+
 //Fetching a single user
 exports.getUser = async (req, res, next) => {
     const userId = req.params.userId;
     try {
         const user = await User.findById(userId);
         if (!user) {
-            const error = new Error('Could not find User.');
-            error.statusCode = 404;
-            throw error;
+            return res
+                .status(404)
+                .json({
+                    message: 'User Not Found'
+                })
+
         }
-        res.status(200).json({ message: 'User fetched.', user: user });
+        res
+            .status(200)
+            .json({
+                message: 'User fetched!', user: user
+            });
     } catch (err) {
         if (!err.statusCode) {
             err.statusCode = 500;
@@ -80,9 +151,18 @@ exports.getUser = async (req, res, next) => {
 
 //Updating the user
 exports.updateUser = async (req, res, next) => {
+    if (req.role !== "admin") {
+        const error = new Error('Updating privileges are granted only admins');
+        error.statusCode = 422;
+        throw error;
+    }
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        return res.status(422).json({message :errors.array()[0].msg})
+        return res
+            .status(422)
+            .json({
+                message: errors.array()[0].msg
+            })
     }
     const userId = req.params.userId;
     const name = req.body.name
@@ -95,18 +175,22 @@ exports.updateUser = async (req, res, next) => {
         const hashedPw = await bcrypt.hash(password, 12);
         const user = await User.findById(userId);
         if (!user) {
-            const error = new Error('Could not find User.');
-            error.statusCode = 404;
-            throw error;
+            return res
+                .status(404)
+                .json({ message: 'User Not Found' })
         }
         user.name = name;
-        user.email = email;
+
         user.password = hashedPw;
         user.role = role;
         user.userAccess = userAccess;
         user.fav = fav;
         const result = await user.save();
-        res.status(200).json({ message: 'User Info updated!', user: result });
+        res
+            .status(200)
+            .json({
+                message: 'User Info updated!', user: result
+            });
     }
     catch (err) {
         if (!err.statusCode) {
@@ -119,12 +203,27 @@ exports.updateUser = async (req, res, next) => {
 //User Deleting
 exports.deleteUser = async (req, res, next) => {
     const userId = req.params.userId;
+
+    if (req.role !== "admin") {
+        const error = new Error('Deleting privileges are granted only admins');
+        error.statusCode = 422;
+        throw error;
+    }
     try {
         const user = await User.findByIdAndRemove(userId);
         if (!user) {
-            return res.status(404).json({ message: 'User not found ' });
+            return res
+                .status(404)
+                .json({
+                    message: 'User not found '
+                });
         }
-        res.status(201).json({ message: "User deleted successfully", name: user.userName });
+        res
+            .status(201)
+            .json({
+                message: "User deleted successfully!",
+                name: user.userName
+            });
     }
     catch (err) {
         if (!err.statusCode) {
@@ -132,5 +231,92 @@ exports.deleteUser = async (req, res, next) => {
         }
         next(err);
     }
-
 }
+
+//fetching fav
+exports.getFav = async (req, res, next) => {
+    const user = await User.findById(req.userId);
+    user
+        .populate('fav.books.bookId')
+
+        .then(user => {
+            const books = user.fav.books;
+            res
+                .status(200)
+                .json({
+                    books: books
+                });
+        })
+        .catch(err => {
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
+};
+
+//adding book to fav
+exports.postFav = async (req, res, next) => {
+    const user = await User.findById(req.userId);
+    const bookId = req.params.bookId;
+    Book.findById(bookId)
+        .then(book => {
+            const dupfavBook = user.fav.books.find(currentBook => {
+                return currentBook.bookId.toString() === book._id.toString();
+            });
+            console.log(dupfavBook);
+            if (dupfavBook) {
+                return res
+                    .status(422)
+                    .json({
+                        message: 'Book already in fav'
+                    })
+            }
+            const updatedFavBooks = [...user.fav.books];
+            updatedFavBooks.push({
+                bookId: book._id,
+            });
+            const updatedFav = {
+                books: updatedFavBooks
+            };
+            user.fav = updatedFav;
+            console.log(user.fav)
+            return user.save();
+
+        }).then(result => {
+            console.log(result);
+            return res
+                .status(201)
+                .json({
+                    message: 'Added book to fav!'
+                })
+        })
+        .catch(err => {
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
+};
+
+//deleting book from fav
+exports.postFavDeleteBook = async (req, res, next) => {
+    const user = await User.findById(req.userId);
+    const bookId = req.params.bookId;
+
+    const updatedFavBooks = user.fav.books.filter(item => {
+        return item.bookId.toString() !== bookId.toString();
+    });
+    user.fav.books = updatedFavBooks;
+    return user.save()
+        .then(result => {
+            return res
+                .status(201)
+                .json({
+                    message: 'Book removed from favorites!'
+                })
+        })
+        .catch(err => {
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
+};
